+++ packages/react-magma-dom/src/components/FormGroup/index.tsx
@@ -1,14 +1,14 @@
 import * as React from 'react';
-import { HiddenStyles } from '../UtilityStyles';
+import { HiddenStyles } from '../../utils/UtilityStyles';
 import styled from '../../theme/styled';
-import { omit, useGenerateId } from '../utils';
+import { omit, useGenerateId } from '../../utils';
 
 export interface FormGroupProps extends React.HTMLAttributes<HTMLDivElement> {
   containerStyle?: React.CSSProperties;
   isTextVisuallyHidden?: boolean;
   labelledById?: string;
   labelStyle?: React.CSSProperties;
-  labelText?: string;
+  labelText?: React.ReactNode;
   testId?: string;
 }