+++ packages/react-magma-dom/src/components/LinkButton/index.tsx
@@ -34,7 +34,7 @@ const LinkButtonComponent = styled.button<LinkButtonProps>`
       outline: 2px dotted
         ${props =>
           props.isInverse
-            ? props.theme.colors.neutral08
+            ? props.theme.colors.focusInverse
             : props.theme.colors.focus};
       outline-offset: 3px;
     }