+++ packages/react-magma-dom/src/components/Tooltip/index.tsx
@@ -2,7 +2,7 @@ import * as React from 'react';
 import styled from '../../theme/styled';
 import { css } from '@emotion/core';
 import { ThemeContext } from '../../theme/ThemeContext';
-import { useGenerateId } from '../utils';
+import { useGenerateId } from '../../utils';
 
 export enum EnumTooltipPosition {
   bottom = 'bottom',
@@ -13,11 +13,13 @@ export enum EnumTooltipPosition {
 
 export interface TooltipProps extends React.HTMLAttributes<HTMLDivElement> {
   children: React.ReactElement;
+  containerStyle?: React.CSSProperties;
   content: React.ReactNode;
   isInverse?: boolean;
   position?: EnumTooltipPosition;
   testId?: string;
   ref?: React.Ref<HTMLDivElement>;
+  tooltipStyle?: React.CSSProperties;
 }
 
 export interface ITooltipState {
@@ -239,10 +241,12 @@ export const Tooltip: React.FunctionComponent<TooltipProps> = React.forwardRef(
     const {
       children,
       content,
+      containerStyle,
       id: defaultId,
       isInverse,
       position,
       testId,
+      tooltipStyle,
       ...other
     } = props;
 
@@ -264,10 +268,9 @@ export const Tooltip: React.FunctionComponent<TooltipProps> = React.forwardRef(
     });
 
     return (
-      <ToolTipContainer data-testid={testId}>
+      <ToolTipContainer {...other} data-testid={testId} style={containerStyle}>
         {tooltipTrigger}
         <StyledTooltip
-          {...other}
           id={id}
           position={position ? position : EnumTooltipPosition.top}
           role="tooltip"
@@ -276,6 +279,7 @@ export const Tooltip: React.FunctionComponent<TooltipProps> = React.forwardRef(
           <StyledTooltipInner
             isInverse={isInverse}
             position={position ? position : EnumTooltipPosition.top}
+            style={tooltipStyle}
             theme={theme}
           >
             {content}