+++ packages/react-magma-dom/src/index.ts
@@ -319,4 +319,4 @@ export { WlUseItIcon } from './components/Icon/types/WlUseItIcon';
 export { WrenchIcon } from './components/Icon/types/WrenchIcon';
 export { YoutubeIcon } from './components/Icon/types/YoutubeIcon';
 
-export { useMediaQuery } from './components/utils';
+export { useMediaQuery } from './utils';