+++ packages/react-magma-dom/src/theme/GlobalStyles/index.tsx
@@ -14,7 +14,7 @@ function getStyles(theme) {
 
     *:focus {
       outline: 2px dotted ${theme.colors.focus};
-      outline-offset: 3px;
+      outline-offset: 4px;
     }
 
     html {